Use existing WgpuNative class
